feat: Implement core FSM logic and basic store operations (#89)
* feat: Implement core FSM logic and basic store operations

This commit introduces the core Finite State Machine (FSM) logic for the Raft-based store, enabling replicated state changes and snapshotting.

Key implementations:
- `Store.fsmApply`: Handles `COMMAND_TYPE_EXECUTE` for "SET key value" and "DELETE key" string commands, applying them to the BadgerDB. Also handles `COMMAND_TYPE_NOOP`. Other command types return `ErrNotImplemented` or a basic response.
- `snapshot.Snapshot.Persist`: Implements snapshot data persistence to a `raft.SnapshotSink`, including writing FSM index/term and the DB data stream.
- `Store.fsmSnapshot`: Creates snapshots of the BadgerDB, including FSM index/term metadata, and streams it via an `io.Pipe` to be consumed by Raft.
- `Store.fsmRestore`: Restores the store's state from a snapshot, including FSM index/term and BadgerDB data. Handles wiping the old DB directory and reloading into a new DB instance.
- `Store.Query` (Placeholder): Basic implementation for direct "GET key" reads from BadgerDB (does not go through Raft).
- Unit Tests: Added `internal/store/store_test.go` with tests for `fsmApply` (SET, DELETE, NOOP, invalid/unimplemented commands), snapshot creation/persistence, and snapshot restoration.

This provides the foundational mechanisms for distributed state management and data persistence in the store. Further work will be needed for more complex query capabilities through Raft and implementing other pending store functionalities.

* feat: handle some race conditions and add new features

---------

+++ internal/partitioner/partition.go
@@ -10,7 +10,7 @@ import (
 
 type Partitoner[T any] struct {
 	// Number of partitions
-	partitions int
+	partitions uint
 	// Hashing function
 	hashFn func(T) (uint64, error)
 	// Buffer size
@@ -28,7 +28,7 @@ type PartitonerOption[T any] func(*Partitoner[T])
 func (p *Partitoner[T]) PartitionData(dataChannel <-chan T) []chan T {
 	partitionedChannels := make([]chan T, p.partitions)
 
-	for i := 0; i < p.partitions; i++ {
+	for i := uint(0); i < p.partitions; i++ {
 		partitionedChannels[i] = make(chan T)
 	}
 
@@ -82,7 +82,7 @@ func WithContext[T any](ctx context.Context) PartitonerOption[T] {
 }
 
 // Partitoner factory function
-func NewPartitoner[T any](partitions int, hashFn func(T) (uint64, error), opts ...PartitonerOption[T]) *Partitoner[T] {
+func NewPartitoner[T any](partitions uint, hashFn func(T) (uint64, error), opts ...PartitonerOption[T]) *Partitoner[T] {
 
 	if hashFn == nil {
 		// TODO: write code to default to round robin
